working on fixing websocket implementation as there are some bugs
